media: platform: tegra: fix checking unsigned val
It fixes checking negative value of an unsigned variable.
"of_get_named_gpio_flags" returns signed int value but
it's stored in unsigned int variable.

Type casted a u32 variable by int to check it's value
is negative.

Coverity ID: 18041

Bug 200116059

Reviewed-on: http://git-master/r/1000774
GVS: Gerrit_Virtual_Submit

+++ drivers/media/platform/tegra/auto/isc_mgr.c
@@ -1,7 +1,7 @@
 /*
  * isc manager.
  *
- * Copyright (c) 2015, NVIDIA Corporation. All Rights Reserved.
+ * Copyright (c) 2015-2016, NVIDIA Corporation. All Rights Reserved.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms and conditions of the GNU General Public License,
@@ -528,7 +528,7 @@ static int isc_mgr_of_get_grp_gpio(
 
 	for (i = 0; (i < num) && (i < size); i++) {
 		grp[i] = of_get_named_gpio_flags(np, name, i, &flags[i]);
-		if (grp[i] < 0) {
+		if ((int)grp[i] < 0) {
 			dev_err(dev, "%s: gpio[%d] invalid\n", __func__, i);
 			return -EINVAL;
 		}
